Transpile, but do not bundle, `abcui.js`
Anybody using our library will be using a build system like Webpack,
so we don't need to support running `abcui.js` as a script.

Instead, we need to ensure that we are shipping `abcui.js` as either a clean
commonjs or es6 library. We don't want to bundle it, since it's a library.