web: Hack the entry point not to crash under node
The post-robot library currently crashes if somebody imports the edge-login-web library in a node.js context. This is because post-robot only wants to run in a browser context.

To avoid crashing our partner's unit tests, we intercept the edge-login-web entry point and replace it with a dummy entry point if the `window` object is missing.