web: Work around bug in Lerna's publish script
Running Rollup and Parcel.js in parallel makes a lot of sense, but somehow `npm-run-all -p` and Lerna aren't getting along correctly. So, switch to running the tools sequentially, which seems to solve the problem.

This fix actually worked the first time (v0.2.7), but I didn't refresh properly to realize the files were there.