rn: Wrap react-redux connect with stronger types
This wrapper knows about our specific `RootState` and `Dispatch` types, so those will always be correct.

Also, both Flow and Typescript have a hard time inferring the types for connect. Flow < 0.89 tends to give up and return `any`. Both Flow >= 0.89 and Typescript tend to ask for more information. So, we provide that information here. This is valid for any version, and actually cleans up the bogus `any` inference we are getting with the current version.

See here for more information:

https://react-redux.js.org/using-react-redux/static-typing