Prevent redundant vouchers from appearing
If the background voucher-checking task happens at the same time an OTP backup modal login, it might produce an extra voucher on the server. This will bug the user to approve their own device, which is confusing. So, pause the background voucher-checking task as long as the OTP backup modal is visible.