Remove workflow-based scene transition design
The login UI design uses workflows, which are composed of one or more
scenes that are indexed by a currentSceneIndex. The view transitions
from scene to scene by incrementing the currentSceneIndex with
ReduxAction WORKFLOW_NEXT. This design of blindly transitioning to the
next scene in the workflow allows for scenarios in which scenes may be
skipped when multiple WORKFLOW_NEXT actions are made. Most workflows
consist only of a single scene.

Replacing the workflow-based grouped scene transition design with a
flat hierarchy of scenes which are explicitly called during each
transition guards against the possibility of unintentionally
transitioning past the intended scene.